Revert "💄 Adjust option items to have same look as combobox"
This reverts commit 50f08c1a2dc6465d4e01ae05d0672c0b6a4bece7.
