+++ src/components/Inputs/Select/OptionDrawer.tsx
@@ -7,15 +7,66 @@ import {
   useState,
 } from 'react';
 
-import { Button, Checkbox } from '@equinor/eds-core-react';
+import { Checkbox, Icon } from '@equinor/eds-core-react';
 import { arrow_drop_down, arrow_drop_up } from '@equinor/eds-icons';
+import { tokens } from '@equinor/eds-tokens';
 
-import {
-  StyledIcon,
-  StyledOption,
-  StyledOptionItem,
-  StyledOptionWrapper,
-} from './OptionDrawer.styles';
+import styled, { css, keyframes } from 'styled-components';
+
+const { colors } = tokens;
+
+interface StyledOptionProps {
+  $section: number;
+  $animationActive?: boolean;
+}
+
+const animateToggle = keyframes`
+  0% {
+    opacity: 1;
+  }
+  90% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 0;
+    display: none;
+  }
+`;
+
+const StyledOptionWrapper = styled.div<StyledOptionProps>`
+  margin-left: ${({ $section }) => ($section > 0 ? '22px' : '')};
+  border-left: ${({ $section }) =>
+    $section > 0 ? '1px solid ' + colors.ui.background__medium.rgba : ''};
+  opacity: 1;
+  color: ${colors.text.static_icons__default.rgba};
+  animation: ${({ $animationActive }) =>
+    $animationActive
+      ? css`
+          ${animateToggle} 400ms ease-in
+        `
+      : 'none'};
+`;
+
+const StyledOption = styled.div<StyledOptionProps>`
+  position: relative;
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  transition: background-color 0.1s ease-in;
+  &:hover {
+    background-color: ${colors.interactive.primary__hover_alt.rgba};
+  }
+
+  svg {
+    fill: ${colors.interactive.primary__resting.rgba};
+  }
+`;
+
+const StyledIcon = styled(Icon)`
+  position: absolute;
+  top: 10px;
+  right: 10px;
+`;
 
 const getAllItems = <T extends { id: string; label: string; children?: T[] }>(
   items: T[]
@@ -87,7 +138,6 @@ export interface OptionDrawerProps<
   animateParent?: Dispatch<SetStateAction<boolean>>;
   openAll?: boolean;
   showIntermediateParent?: boolean;
-  parentHasNestedItems?: boolean;
 }
 
 /**
@@ -107,7 +157,6 @@ const OptionDrawer = <
   animateParent,
   openAll,
   showIntermediateParent = false,
-  parentHasNestedItems = false,
 }: OptionDrawerProps<T>) => {
   const [open, setOpen] = useState(false);
   const [status, setStatus] = useState<StatusType>(
@@ -130,10 +179,26 @@ const OptionDrawer = <
   }, [openAll]);
 
   const handleClick = (e: MouseEvent) => {
-    if (item.children && item.children.length !== 0) {
-      setOpen((o) => !o);
+    const checkboxElement = e.target as SVGPathElement;
+    const checkboxSVGElement = e.target as SVGElement;
+
+    // if element is of type checkbox, do not take any action.
+    if (
+      !(checkboxElement.getAttribute('name') === 'checked') &&
+      !(checkboxElement.getAttribute('name') === 'not-checked') &&
+      !(
+        checkboxSVGElement.children.length > 0 &&
+        checkboxSVGElement.children[0].getAttribute('name') === 'checked'
+      ) &&
+      !(e.target instanceof HTMLInputElement) &&
+      !(e.target instanceof HTMLSpanElement)
+    ) {
+      if (item.children && item.children.length !== 0) {
+        setOpen((o) => !o);
+      } else if (e.target instanceof HTMLDivElement) {
+        handleCheck(e);
+      }
     }
-    handleCheck(e);
   };
 
   const handleToggle = (e: MouseEvent | ChangeEvent) => {
@@ -204,12 +269,6 @@ const OptionDrawer = <
     }
   };
 
-  const handleChevronIconClick = () => {
-    setOpen((state) => !state);
-  };
-
-  const hasChildren = item.children && item.children.length !== 0;
-
   return (
     <StyledOptionWrapper
       key={
@@ -221,25 +280,18 @@ const OptionDrawer = <
       $animationActive={animationActive}
       data-testid={animationActive ? 'animated-' + item.id : item.id}
     >
-      <StyledOptionItem $paddedLeft={parentHasNestedItems && !hasChildren}>
-        {hasChildren && (
-          <Button
-            variant="ghost_icon"
-            onClick={handleChevronIconClick}
-            data-testid="chevron-toggler-option-drawer"
-          >
-            <StyledIcon data={open ? arrow_drop_up : arrow_drop_down} />
-          </Button>
+      <StyledOption $section={section} onClick={handleClick}>
+        <Checkbox
+          indeterminate={status === 'INTERMEDIATE'}
+          checked={status === 'CHECKED'}
+          onChange={(e) => handleCheck(e)}
+          color="secondary"
+        />
+        {item.label}
+        {item.children && item.children.length !== 0 && (
+          <StyledIcon data={open ? arrow_drop_up : arrow_drop_down} />
         )}
-        <StyledOption $section={section} onClick={handleClick}>
-          <Checkbox
-            indeterminate={status === 'INTERMEDIATE'}
-            checked={status === 'CHECKED'}
-            color="secondary"
-          />
-          {item.label}
-        </StyledOption>
-      </StyledOptionItem>
+      </StyledOption>
       {open &&
         item.children?.map((i) => (
           <OptionDrawer<T>
@@ -255,7 +307,6 @@ const OptionDrawer = <
             animateParent={setAnimationActive}
             openAll={openAll}
             showIntermediateParent={showIntermediateParent}
-            parentHasNestedItems
           />
         ))}
     </StyledOptionWrapper>
